Switch page check logic order for readability/future selves
The previous framing of avoiding infinite loops didn't really make the intention clear. The intention here is to check to see if we are already on the page that is unrestricted. It's possible this check should really happen in a different function but let's not change that now since that could affect hooks people are expecting to fire.