Add a check for the special case of a static home page
This is a somewhat less common situation, where the plugin is set up to show the static homepage but nothing else on the site. This became an infinite loop when we switched away from manipulating the `$wp` object in favor of redirecting to a URL in #30, which is more stable for all other pages but, surprise, the special way static front pages are handled did not go over well.

This really needs automated tests.